refactor(swarm_team): move plan response format to types
